+++ src/main/java/com/teamfaceless/facelessjobs/controller/EmpresaSecurityController.java
@@ -29,7 +29,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 @Controller
 @RequestMapping("/empresa")
-public class ControllerEmpresa {
+public class EmpresaSecurityController {
 
 	@Autowired
 	private IEmpresaService iEmpresaService;
@@ -45,38 +45,6 @@ public class ControllerEmpresa {
 	
 	@Autowired
 	private IValidations iValidations;
-	
-	@GetMapping("/pruebas/{idEmpresa}")
-	public String goPruebas(@PathVariable Integer idEmpresa, Model model) {
-		model.addAttribute("provincias", iProvinciaService.findAll());
-		model.addAttribute("sectores", iSectorService.findAll());
-		model.addAttribute("empresa", iEmpresaService.findById(idEmpresa));
-		return "views/empresa/pruebas";
-	}
-	
-	@PostMapping("/pruebas")
-	public String postPruebas(Empresa empresa, Model model, BindingResult result) {
-		if (result.hasErrors()) {
-			model.addAttribute("empresa", empresa);
-			model.addAttribute("provincias", iProvinciaService.findAll());
-			model.addAttribute("sectores", iSectorService.findAll());
-			return "views/empresa/pruebas";
-		}
-		iEmpresaService.modify(empresa);
-		return "redirect:/empresa/listado";
-	}
-	
-	@GetMapping("/perfil")
-	public String goPerfil(Model model) {
-		model.addAttribute("titulo", "Perfil de Empresa");
-		return "views/empresa/perfil";
-	}
-	
-	@GetMapping("/listado")
-	public String goListado(Model model) {
-		model.addAttribute("listEmpresas", iEmpresaService.findAll());
-		return "/views/empresa/listado";
-	}
 	
 	@GetMapping("/registro")
 	public String formRegistro(Model model, EmpresaRegistroDto empresaRegistroDto) {
@@ -118,39 +86,6 @@ public class ControllerEmpresa {
 		return "redirect:/empresa/listado";
 	}
 	
-	@GetMapping("/modificar/{idEmpresa}")
-	public String goModificar(@PathVariable Integer idEmpresa, Model model) {
-		model.addAttribute("empresa", iEmpresaService.findById(idEmpresa));
-		model.addAttribute("empresaregistrodto", new EmpresaRegistroDto());
-		model.addAttribute("provincias", iProvinciaService.findAll());
-		model.addAttribute("sectores", iSectorService.findAll());
-		return "/views/empresa/detalle";
-	}
-	
-	@PostMapping("/modificar")
-	public String modificarEmpresa(Model model, @Valid Empresa empresa, BindingResult result) {
-		if (result.hasErrors()) {
-			model.addAttribute("empresa", empresa);
-			model.addAttribute("provincias", iProvinciaService.findAll());
-			model.addAttribute("sectores", iSectorService.findAll());
-			return "/views/empresa/detalle";
-		}
-		iEmpresaService.modify(empresa);
-		return "redirect:/empresa/listado";
-	}
-
-	@GetMapping("/borrar/{idEmpresa}")
-	public String formBaja(@PathVariable Integer idEmpresa) {
-		iEmpresaService.delete(idEmpresa);
-		return "redirect:/empresa/listado";
-	}
-	
-	@PostMapping("/borrar/{idEmpresa}")
-	public String bajaEmpresa(@PathVariable Integer idEmpresa) {
-		iEmpresaService.delete(idEmpresa);
-		return "redirect:/empresa/listado";
-	}
-	
 	@GetMapping("/login")
 	public String formLogin(Model model,@RequestParam(value = "error",required = false)String error) {
 		model.addAttribute("credencial", new Credencial());
@@ -168,21 +103,6 @@ public class ControllerEmpresa {
 		}
 		//loguear
 		return"views/app/empresa/home";
-	}
-	
-	@GetMapping("/pruebasValidation")
-	private String goPruebasValidation(Model model) {
-		model.addAttribute("empresa", new EmpresaListadoDto());
-		return "/views/empresa/pruebasValidation";
-	}
-	
-	@PostMapping("/pruebasValidation")
-	private String goPruebasValidationReturn(Model model,@Valid EmpresaListadoDto empresaListadoDto, BindingResult result) {
-		if (result.hasErrors()) {
-			model.addAttribute("empresa", empresaListadoDto);
-			return "/views/empresa/pruebasValidation";
-		}
-		return "redirect:/empresa/listado";
 	}
 	
 }
