+++ src/main/resources/templates/views/app/empresa/detalle.html
@@ -1,13 +1,11 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org">
-<head th:replace="plantilla/template :: head"></head>
+<head th:replace="fragments/base :: head"></head>
 <body>
-<!-- 	<header th:replace="plantilla/template :: header"></header> -->
 	<nav th:replace="fragments/base :: navbar"></nav>
 	<h1 class="text-center bg-danger">Modificar Empresa</h1>
 	<h2 th:text="${titulo}">Sutitulo</h2>
 	<hr>
-
 	<main class="row">
 		<div class="col-2"></div>
 		<formulario class="col-8">
