+++ src/main/resources/templates/views/app/empresa/formulario_pruebas.html
@@ -1,8 +1,12 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org">
-<head th:replace="plantilla/template :: head"></head>
+<head th:replace="fragments/base :: head">
+
+</head>
+
+
 <body>
-	<header th:replace="plantilla/template :: header"></header>
+	<nav th:replace="fragments/base :: navbar"></nav>
 	<h1 class="text-center">Soy el formulario de empresas</h1>
 	<h2 th:text="${titulo}">Sutitulo</h2>
 	<hr>
@@ -343,6 +347,7 @@
 			</div>
 		</div>
 	</form>
-	<footer th:replace="plantilla/template :: footer"></footer>
+	<footer th:replace="fragments/base :: footer"></footer>
+  	<th:block th:replace="fragments/base :: script"></th:block>
 </body>
 </html>
