+++ src/main/resources/templates/views/app/empresa/home.html
@@ -1,22 +1,13 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org">
-<head th:replace="plantilla/template-registro-login :: head">
-
-</head>
-
+<head th:replace="fragments/base :: head"></head>
 <body>
-	<header th:replace="plantilla/template-registro-login :: header">
-
-	</header>
-
+	<nav th:replace="fragments/base :: navbar"></nav>
 	<h3>Logueado Empresa</h3>
 
 
-	<footer th:replace="plantilla/template-registro-login :: footer">
-
-	</footer>
-	<div th:replace="plantilla/template-registro-login :: boostrap-js"></div>
-
+  <footer th:replace="fragments/base :: footer"></footer>
+  <th:block th:replace="fragments/base :: script"></th:block>
 </body>
 
 </html>
