+++ src/main/resources/templates/views/app/empresa/listado.html
@@ -1,9 +1,8 @@
 <!DOCTYPE html>
 <html xmlns:th="html://www.thymeleaf.org">
-<head th:replace="plantilla/template :: head" />
+<head th:replace="fragments/base :: head"></head>
 <body>
-  	<nav th:replace="fragments/base :: navbar"></nav>
-<!-- 	<header th:replace="plantilla/template :: header"></header> -->
+	<nav th:replace="fragments/base :: navbar"></nav>
 	<hr>
 	<h1 class="text-center">Listado Empresas</h1>
 	<hr>
