+++ src/main/resources/templates/views/app/empresa/pruebas.html
@@ -1,8 +1,7 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org">
-<head th:replace="plantilla/template :: head"></head>
+<head th:replace="fragments/base :: head"></head>
 <body>
-<!-- 	<header th:replace="plantilla/template :: header"></header> -->
 	<nav th:replace="fragments/base :: navbar"></nav>
 	<h1 class="text-center">Pagina de pruebas</h1>
 	En pruebas
