+++ src/main/resources/templates/views/app/empresa/registroAntonio.html
@@ -1,14 +1,8 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org">
-<head th:replace="plantilla/template-registro-login :: head">
-
-</head>
-
+<head th:replace="fragments/base :: head"></head>
 <body>
-	<header th:replace="plantilla/template-registro-login :: header">
-
-	</header>
-
+	<nav th:replace="fragments/base :: navbar"></nav>
 	<div class="container wrapper">
 
 		<div class="row g-5 justify-content-center p-5">
@@ -92,11 +86,8 @@
 		</div>
 
 	</div>
-	<footer th:replace="plantilla/template-registro-login :: footer">
-
-	</footer>
-	<div th:replace="plantilla/template-registro-login :: boostrap-js"></div>
-
+  	<footer th:replace="fragments/base :: footer"></footer>
+  	<th:block th:replace="fragments/base :: script"></th:block>
 </body>
 
 </html>
